requirements_test.txt
cryptography
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.0
pytest-homeassistant-custom-component>=0.13.0
ruff>=0.11
salus-it600-client==0.4.4
